+++ packages/react-app/src/index.jsx
@@ -5,6 +5,7 @@ import { BrowserRouter } from "react-router-dom";
 import ReactDOM from "react-dom";
 import App from "./App";
 import "./index.css";
+import StateContext from "./context/StateContext";
 
 const themes = {
   dark: `${process.env.PUBLIC_URL}/dark-theme.css`,
@@ -24,7 +25,9 @@ ReactDOM.render(
   <ApolloProvider client={client}>
     <ThemeSwitcherProvider themeMap={themes} defaultTheme={prevTheme || "light"}>
       <BrowserRouter>
-        <App subgraphUri={subgraphUri} />
+        <StateContext>
+          <App subgraphUri={subgraphUri} />
+        </StateContext>
       </BrowserRouter>
     </ThemeSwitcherProvider>
   </ApolloProvider>,
